fix: Apply code fix according to reviewer s recomendations
