+++ stats/pom.xml
@@ -23,6 +23,9 @@
         <maven.compiler.source>21</maven.compiler.source>
         <maven.compiler.target>21</maven.compiler.target>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+        <lombok.version>1.18.34</lombok.version>
+        <jackson.version>2.15.2</jackson.version>
+        <h2.version>2.3.232</h2.version>
     </properties>
 
 </project>
