+++ stats/stats-dto/src/main/java/ru/practicum/ViewStatsDto.java
@@ -2,12 +2,14 @@ package ru.practicum;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.PositiveOrZero;
+import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.experimental.Accessors;
 
 @Getter
 @Setter
+@AllArgsConstructor
 @Accessors(chain = true)
 public class ViewStatsDto {
