--- stats/stats-server/src/main/java/ru/practicum/StatsRepository.java
@@ -1,8 +1,0 @@
-package ru.practicum;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import ru.practicum.model.EndPointHit;
-
-public interface StatsRepository extends JpaRepository<EndPointHit, Integer>, StatsCustomRepository {
-
-}
