--- stats/stats-server/src/main/java/ru/practicum/model/ViewStats.java
@@ -1,18 +1,0 @@
-package ru.practicum.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-@AllArgsConstructor
-public class ViewStats {
-
-  private String app;
-
-  private String uri;
-
-  private Long hits;
-
-}
