+++ stats/stats-server/src/main/java/ru/practicum/repository/StatsRepository.java
@@ -1,0 +1,8 @@
+package ru.practicum.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import ru.practicum.model.EndPointHit;
+
+public interface StatsRepository extends JpaRepository<EndPointHit, Integer>, StatsQueryRepository {
+
+}
