+++ stats/stats-server/src/main/resources/application.properties
@@ -18,7 +18,7 @@ logging.level.org.springframework.transaction=INFO
 
 # PostgreSQL Database Configuration(Default)
 spring.datasource.driverClassName=org.postgresql.Driver
-spring.datasource.url=jdbc:postgresql://localhost:5432/stat_service
+#spring.datasource.url=jdbc:postgresql://localhost:5432/stat_service
 spring.datasource.username=postgres
 spring.datasource.password=admin
-#spring.datasource.url=jdbc:postgresql://localhost:6541/stat_service
+spring.datasource.url=jdbc:postgresql://localhost:6541/stat_service
